+++ src/cmd/cmd_c.ml
@@ -166,7 +166,8 @@ let cmd ~debug ~arch ~property ~testcomp:_ ~workspace ~workers ~opt_lvl
   ~includes ~files ~profiling ~unsafe ~optimize ~no_stop_at_failure ~no_value
   ~no_assert_failure_expression_printing ~deterministic_result_order ~fail_mode
   ~concolic ~eacsl ~solver ~profile ~model_format ~(entry_point : string option)
-  ~invoke_with_symbols ~out_file ~model_out_file : unit Result.t =
+  ~invoke_with_symbols ~out_file ~model_out_file ~with_breadcrumbs :
+  unit Result.t =
   let* workspace =
     match workspace with
     | Some path -> Ok path
@@ -189,3 +190,4 @@ let cmd ~debug ~arch ~property ~testcomp:_ ~workspace ~workers ~opt_lvl
     ~no_stop_at_failure ~no_value ~no_assert_failure_expression_printing
     ~deterministic_result_order ~fail_mode ~workspace ~solver ~files ~profile
     ~model_format ~entry_point ~invoke_with_symbols ~model_out_file
+    ~with_breadcrumbs