+++ src/cmd/cmd_conc.mli
@@ -23,4 +23,5 @@ val cmd :
   -> entry_point:string option
   -> invoke_with_symbols:bool
   -> model_out_file:Fpath.t option
+  -> with_breadcrumbs:bool
   -> unit Result.t