+++ src/libc/include/owi.h
@@ -27,6 +27,8 @@ __attribute__((import_module("symbolic"), import_name("range_symbol"))) int
 owi_range(int lo, int hi);
 __attribute__((import_module("symbolic"), import_name("print_char"))) void
 owi_print_char(int);
+__attribute__((import_module("symbolic"), import_name("label"))) void
+owi_label(int id, char * name);
 __attribute__((import_module("symbolic"), import_name("bool_symbol")))
 #ifdef __cplusplus
   bool